feat: use completed and dismiss states
