+++ src/data/reducers/enterpriseCustomerAdmin.js
@@ -2,7 +2,6 @@ import {
   FETCH_ENTERPRISE_CUSTOMER_ADMIN_REQUEST,
   FETCH_ENTERPRISE_CUSTOMER_ADMIN_SUCCESS,
   FETCH_ENTERPRISE_CUSTOMER_ADMIN_FAILURE,
-  UPDATE_ENTERPRISE_CUSTOMER_ADMIN,
 } from '../constants/enterpriseCustomerAdmin';
 
 const initialState = {
